Moved renumber to AutoFormat
