+++ app/src/main/java/net/gsantner/markor/format/markdown/MarkdownTextActions.java
@@ -11,7 +11,6 @@ package net.gsantner.markor.format.markdown;
 
 import android.app.Activity;
 import android.support.annotation.StringRes;
-import android.text.Editable;
 import android.view.HapticFeedbackConstants;
 import android.view.KeyEvent;
 import android.view.View;
@@ -30,18 +29,17 @@ import net.gsantner.opoc.util.StringUtils;
 import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.List;
-import java.util.regex.Matcher;
 import java.util.regex.Pattern;
 
 public class MarkdownTextActions extends TextActions {
 
-    private static final Pattern PREFIX_ORDERED_LIST = Pattern.compile("^(\\s*)((\\d+)(\\.|\\))(\\s+))");
-    private static final Pattern PREFIX_ATX_HEADING = Pattern.compile("^(\\s{0,3})(#{1,6}\\s)");
-    private static final Pattern PREFIX_QUOTE = Pattern.compile("^(>\\s)");
-    private static final Pattern PREFIX_CHECKED_LIST = Pattern.compile("^(\\s*)((-|\\*|\\+)\\s\\[(x|X)]\\s)");
-    private static final Pattern PREFIX_UNCHECKED_LIST = Pattern.compile("^(\\s*)((-|\\*|\\+)\\s\\[\\s]\\s)");
-    private static final Pattern PREFIX_UNORDERED_LIST = Pattern.compile("^(\\s*)((-|\\*|\\+)\\s)");
-    private static final Pattern PREFIX_LEADING_SPACE = Pattern.compile("^(\\s*)");
+    public static final Pattern PREFIX_ORDERED_LIST = Pattern.compile("^(\\s*)((\\d+)(\\.|\\))(\\s+))");
+    public static final Pattern PREFIX_ATX_HEADING = Pattern.compile("^(\\s{0,3})(#{1,6}\\s)");
+    public static final Pattern PREFIX_QUOTE = Pattern.compile("^(>\\s)");
+    public static final Pattern PREFIX_CHECKED_LIST = Pattern.compile("^(\\s*)((-|\\*|\\+)\\s\\[(x|X)]\\s)");
+    public static final Pattern PREFIX_UNCHECKED_LIST = Pattern.compile("^(\\s*)((-|\\*|\\+)\\s\\[\\s]\\s)");
+    public static final Pattern PREFIX_UNORDERED_LIST = Pattern.compile("^(\\s*)((-|\\*|\\+)\\s)");
+    public static final Pattern PREFIX_LEADING_SPACE = Pattern.compile("^(\\s*)");
 
     private static final Pattern[] PREFIX_PATTERNS = {
             PREFIX_ORDERED_LIST,
@@ -154,7 +152,7 @@ public class MarkdownTextActions extends TextActions {
                 case R.string.tmaid_common_ordered_list_number: {
                     runPrefixReplaceAction(PREFIX_ORDERED_LIST, "$11. ", "$1");
                     if (_appSettings.isMarkdownAutoUpdateList()) {
-                        renumberOrderedList(_hlEditor.getText(), StringUtils.getSelection(_hlEditor)[0]);
+                        MarkdownAutoFormat.renumberOrderedList(_hlEditor.getText(), StringUtils.getSelection(_hlEditor)[0]);
                     }
                     return true;
                 }
@@ -237,7 +235,7 @@ public class MarkdownTextActions extends TextActions {
                     return true;
                 }
                 case R.string.tmaid_common_ordered_list_number: {
-                    renumberOrderedList(_hlEditor.getText(), StringUtils.getSelection(_hlEditor)[0]);
+                    MarkdownAutoFormat.renumberOrderedList(_hlEditor.getText(), StringUtils.getSelection(_hlEditor)[0]);
                 }
             }
             return false;
@@ -312,184 +310,5 @@ public class MarkdownTextActions extends TextActions {
         }
 
         runRegexReplaceAction(patterns);
-    }
-
-
-    private static class ListLine {
-
-        protected static final int INDENT_DELTA = 2;
-
-        public final int lineStart, lineEnd;
-        public final String line;
-        public final int indent;
-        public final boolean isEmpty;
-
-        public ListLine(CharSequence text, int position) {
-
-            lineStart = StringUtils.getLineStart(text, position);
-            lineEnd = StringUtils.getLineEnd(text, position);
-            line = text.subSequence(lineStart, lineEnd).toString();
-            indent = StringUtils.getNextNonWhitespace(text, lineStart) - lineStart;
-            isEmpty = (lineEnd - lineStart) == indent;
-        }
-
-        public boolean isChild(final ListLine line) {
-            return line.isEmpty || line.indent > (indent + INDENT_DELTA);
-        }
-
-        public boolean isParent(final ListLine line) {
-            return !line.isEmpty && line.indent < (indent - INDENT_DELTA);
-        }
-    }
-
-    /**
-     * Class to parse a line of text and extract useful information
-     */
-    private static class OrderedListLine extends ListLine {
-        private static final int VALUE_GROUP = 3;
-        private static final int DELIM_GROUP = 4;
-
-        public final boolean isOrderedList;
-        public final char delimiter;
-        public final int numStart, numEnd;
-        public final int value;
-
-        public OrderedListLine(CharSequence text, int position) {
-            super(text, position);
-
-            final Matcher match = PREFIX_ORDERED_LIST.matcher(line);
-            isOrderedList = match.find();
-            if (isOrderedList) {
-                delimiter = match.group(DELIM_GROUP).charAt(0);
-                numStart = match.start(VALUE_GROUP) + lineStart;
-                numEnd = match.end(VALUE_GROUP) + lineStart;
-                value = Integer.parseInt(match.group(VALUE_GROUP));
-            } else {
-                numStart = numEnd = value = -1;
-                delimiter = 0;
-            }
-        }
-
-        public boolean isMatchingList(final OrderedListLine line) {
-            final boolean bothOrderedlists = isOrderedList && line.isOrderedList;
-            final boolean sameIndent = Math.abs(indent - line.indent) <= ListLine.INDENT_DELTA;
-            final boolean sameDelimiter = delimiter == line.delimiter;
-            return bothOrderedlists && sameIndent && sameDelimiter;
-        }
-    }
-
-    private static class UnOrderedListLine extends ListLine {
-        private static final int CHECK_GROUP = 4;
-        private static final int LIST_LEADER_GROUP = 3;
-        private static final int FULL_GROUP = 2;
-
-        public final boolean isUnorderedList;
-        public final boolean isCheckboxList;
-        public final boolean isChecked;
-        public final char checkChar;
-        public final char listChar;
-        public final int groupStart, groupEnd;
-
-        public UnOrderedListLine(CharSequence text, int position) {
-            super(text, position);
-
-            final Matcher ucMatch = PREFIX_UNCHECKED_LIST.matcher(line);
-            final Matcher cMatch = PREFIX_CHECKED_LIST.matcher(line);
-            final Matcher uMatch = PREFIX_UNORDERED_LIST.matcher(line);
-
-            isUnorderedList = uMatch.find(); // Will also catch other unordered list types
-            isCheckboxList = ucMatch.find() || cMatch.find();
-            isChecked = cMatch.find() && !ucMatch.find();
-
-            if (isChecked) {
-                checkChar = cMatch.group(CHECK_GROUP).charAt(0);
-            } else {
-                checkChar = 0;
-            }
-
-            if (isUnorderedList) {
-                listChar = uMatch.group(LIST_LEADER_GROUP).charAt(0);
-                Matcher match = isCheckboxList ? (isChecked ? cMatch : ucMatch) : uMatch;
-                groupStart = lineStart + match.start(FULL_GROUP);
-                groupEnd = lineStart + match.end(FULL_GROUP);
-            } else {
-                listChar = 0;
-                groupStart = groupEnd = -1;
-            }
-        }
-    }
-
-    /**
-     * Walks to the top of the current list at the current level
-     * <p>
-     * This function will not walk to parent levels!
-     *
-     * @param searchStart position to start search at
-     * @return OrderedLine corresponding to top of the list
-     */
-    private static OrderedListLine getOrderedListStart(Editable text, final int searchStart) {
-
-        int position = Math.max(Math.min(searchStart, text.length() - 1), 0);
-
-        OrderedListLine line, listStart = null, startLine = null;
-
-        do {
-            line = new OrderedListLine(text, position);
-
-            if (startLine == null) {
-                startLine = line;
-                if (!startLine.isOrderedList) {
-                    break;
-                }
-            }
-
-            if (startLine.isMatchingList(line)) {
-                listStart = line;
-            }
-
-            position = line.lineStart - 1;
-
-        } while (position > 0 && (startLine.isMatchingList(line) || startLine.isChild(line)));
-
-        return listStart == null ? line : listStart;
-    }
-
-
-    /**
-     * This function will first walk up to the top of the current list level
-     * and then walk down to the end of the current list level.
-     * <p>
-     * Sub-lists and other children will be skipped.
-     */
-    public static void renumberOrderedList(Editable text, int cursorPosition) {
-
-        // Top of list
-        final OrderedListLine firstLine = getOrderedListStart(text, cursorPosition);
-
-        if (firstLine.isOrderedList && firstLine.lineEnd < text.length()) {
-            int number = firstLine.value;
-
-            int position = firstLine.lineEnd + 1;
-            while (position >= 0 && position < text.length()) {
-
-                final OrderedListLine line = new OrderedListLine(text, position);
-
-                if (firstLine.isMatchingList(line)) {
-                    number += 1;
-                    if (line.value != number) {
-                        String newNum = Integer.toString(number);
-                        text.replace(line.numStart, line.numEnd, newNum);
-                        final int lenDiff = newNum.length() - (line.numEnd - line.numStart);
-                        position = line.lineEnd + lenDiff + 1;
-                    } else {
-                        position = line.lineEnd + 1;
-                    }
-                } else if (firstLine.isChild(line)) {
-                    position = line.lineEnd + 1;
-                } else {
-                    break;
-                }
-            }
-        }
     }
 }
